+++ spec/eve_online/esi/character_bookmarks_spec.rb
@@ -17,7 +17,40 @@ describe EveOnline::ESI::CharacterBookmarks do
     its(:character_id) { should eq(12_345_678) }
   end
 
-  # describe '#bookmarks' # TODO: write this
+  describe '#bookmarks' do
+    let(:bookmark) { double }
+
+    let(:response) do
+      [
+        {
+          bookmark_id: 4,
+          created: '2012-07-09T22:38:31Z',
+          label: 'Stargate',
+          notes: 'This is a stargate',
+          location_id: 30_003_430,
+          creator_id: 2_112_625_428
+        }
+      ]
+    end
+
+    before do
+      #
+      # subject.response # => [{"bookmark_id"=>4, "created"=>"2012-07-09T22:38:31Z", "label"=>"Stargate", "notes"=>"This is a stargate", "location_id"=>30003430, "creator_id"=>2112625428}]
+      #
+      expect(subject).to receive(:response).and_return(response)
+    end
+
+    before do
+      #
+      # EveOnline::ESI::Models::Bookmark.new(response.first) # => bookmark
+      #
+      expect(EveOnline::ESI::Models::Bookmark).to receive(:new).with(response.first).and_return(bookmark)
+    end
+
+    specify { expect(subject.bookmarks).to eq([bookmark]) }
+
+    specify { expect { subject.bookmarks }.to change { subject.instance_variable_defined?(:@_memoized_bookmarks) }.from(false).to(true) }
+  end
 
   describe '#scope' do
     specify { expect(subject.scope).to eq('esi-bookmarks.read_character_bookmarks.v1') }